+++ tests/Silverback.Integration.Tests.E2E/Mqtt/EncryptionTests.cs
@@ -23,8 +23,13 @@ namespace Silverback.Tests.Integration.E2E.Mqtt
         private static readonly byte[] AesEncryptionKey =
         {
             0x0d, 0x0e, 0x0f, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c,
-            0x1d, 0x1e,
-            0x1f, 0x20, 0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c
+            0x1d, 0x1e, 0x1f, 0x20, 0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c
+        };
+
+        private static readonly byte[] AesEncryptionKey2 =
+        {
+            0x2d, 0x2e, 0x0f, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c,
+            0x1d, 0x1e, 0x1f, 0x20, 0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c
         };
 
         public EncryptionTests(ITestOutputHelper testOutputHelper)
@@ -70,12 +75,80 @@ namespace Silverback.Tests.Integration.E2E.Mqtt
             Helper.Spy.OutboundEnvelopes.Should().HaveCount(2);
             Helper.Spy.OutboundEnvelopes[0].RawMessage.Should().BeOfType<SymmetricEncryptStream>();
             Helper.Spy.OutboundEnvelopes[1].RawMessage.Should().BeOfType<SymmetricEncryptStream>();
+            Helper.Spy.OutboundEnvelopes.ForEach(
+                envelope =>
+                    envelope.Headers.GetValue(DefaultMessageHeaders.EncryptionKeyId).Should().BeNull());
 
             Helper.Spy.InboundEnvelopes.Should().HaveCount(2);
             Helper.Spy.InboundEnvelopes[0].Message.Should().BeEquivalentTo(message1);
             Helper.Spy.InboundEnvelopes[1].Message.Should().BeEquivalentTo(message2);
+            Helper.Spy.InboundEnvelopes.ForEach(
+                envelope =>
+                    envelope.Headers.GetValue(DefaultMessageHeaders.EncryptionKeyId).Should().BeNull());
+        }
 
-            Helper.GetClientSession("e2e-test").PendingMessagesCount.Should().Be(0);
+        [Fact]
+        public async Task Encryption_SimpleMessagesWithKeyRotation_EncryptedAndDecrypted()
+        {
+            var message1 = new TestEventOne { Content = "Message 1" };
+            var message2 = new TestEventTwo { Content = "Message 2" };
+            const string keyIdentifier1 = "my-encryption-key-1";
+            const string keyIdentifier2 = "my-encryption-key-2";
+
+            Host.ConfigureServices(
+                    services => services
+                        .AddLogging()
+                        .AddSilverback()
+                        .UseModel()
+                        .WithConnectionToMessageBroker(options => options.AddMockedMqtt())
+                        .AddMqttEndpoints(
+                            endpoints => endpoints
+                                .Configure(
+                                    config => config
+                                        .WithClientId("e2e-test")
+                                        .ConnectViaTcp("e2e-mqtt-broker"))
+                                .AddOutbound<TestEventOne>(
+                                    endpoint => endpoint
+                                        .ProduceTo(DefaultTopicName)
+                                        .EncryptUsingAes(AesEncryptionKey, keyIdentifier1))
+                                .AddOutbound<TestEventTwo>(
+                                    endpoint => endpoint
+                                        .ProduceTo(DefaultTopicName)
+                                        .EncryptUsingAes(AesEncryptionKey2, keyIdentifier2))
+                                .AddInbound(
+                                    endpoint => endpoint
+                                        .ConsumeFrom(DefaultTopicName)
+                                        .DecryptUsingAes(
+                                            keyIdentifier =>
+                                            {
+                                                switch (keyIdentifier)
+                                                {
+                                                    case keyIdentifier1:
+                                                        return AesEncryptionKey;
+                                                    default:
+                                                        return AesEncryptionKey2;
+                                                }
+                                            })))
+                        .AddIntegrationSpyAndSubscriber())
+                .Run();
+
+            var publisher = Host.ScopedServiceProvider.GetRequiredService<IEventPublisher>();
+            await publisher.PublishAsync(message1);
+            await publisher.PublishAsync(message2);
+
+            await Helper.WaitUntilAllMessagesAreConsumedAsync();
+
+            Helper.Spy.OutboundEnvelopes.Should().HaveCount(2);
+            Helper.Spy.OutboundEnvelopes[0].RawMessage.Should().BeOfType<SymmetricEncryptStream>();
+            Helper.Spy.OutboundEnvelopes[0].Headers.GetValue(DefaultMessageHeaders.EncryptionKeyId).Should().Be(keyIdentifier1);
+            Helper.Spy.OutboundEnvelopes[1].RawMessage.Should().BeOfType<SymmetricEncryptStream>();
+            Helper.Spy.OutboundEnvelopes[1].Headers.GetValue(DefaultMessageHeaders.EncryptionKeyId).Should().Be(keyIdentifier2);
+
+            Helper.Spy.InboundEnvelopes.Should().HaveCount(2);
+            Helper.Spy.InboundEnvelopes[0].Message.Should().BeEquivalentTo(message1);
+            Helper.Spy.InboundEnvelopes[0].Headers.GetValue(DefaultMessageHeaders.EncryptionKeyId).Should().Be(keyIdentifier1);
+            Helper.Spy.InboundEnvelopes[1].Message.Should().BeEquivalentTo(message2);
+            Helper.Spy.InboundEnvelopes[1].Headers.GetValue(DefaultMessageHeaders.EncryptionKeyId).Should().Be(keyIdentifier2);
         }
 
         [Fact]