+++ tests/Silverback.Integration.Tests/Messaging/Configuration/ConsumerEndpointBuilderDecryptUsingExtensionsTests.cs
@@ -19,11 +19,33 @@ namespace Silverback.Tests.Integration.Messaging.Configuration
             var key = new byte[] { 0x01, 0x02, 0x03, 0x04, 0x05 };
             var iv = new byte[] { 0x11, 0x12, 0x13, 0x14, 0x15 };
 
-            var endpoint = builder.DecryptUsingAes(key, iv).Build();
+            TestConsumerEndpoint endpoint = builder.DecryptUsingAes(key, iv).Build();
 
-            endpoint.Encryption.Should().BeOfType<SymmetricEncryptionSettings>();
-            endpoint.Encryption.As<SymmetricEncryptionSettings>().Key.Should().BeSameAs(key);
-            endpoint.Encryption.As<SymmetricEncryptionSettings>().InitializationVector.Should().BeSameAs(iv);
+            endpoint.Encryption.Should().BeOfType<SymmetricDecryptionSettings>();
+            endpoint.Encryption.As<SymmetricDecryptionSettings>().Key.Should().BeSameAs(key);
+            endpoint.Encryption.As<SymmetricDecryptionSettings>().InitializationVector.Should().BeSameAs(iv);
+        }
+
+        [Fact]
+        public void DecryptUsingAes_WithKeyProvider_EncryptionSettingsSet()
+        {
+            var builder = new TestConsumerEndpointBuilder();
+
+            var key = new byte[] { 0x01, 0x02, 0x03, 0x04, 0x05 };
+            var iv = new byte[] { 0x11, 0x12, 0x13, 0x14, 0x15 };
+
+            byte[] DecryptionKeyCallback(string keyIdentifier)
+            {
+                return key;
+            }
+
+            TestConsumerEndpoint endpoint = builder.DecryptUsingAes(DecryptionKeyCallback, iv).Build();
+
+            endpoint.Encryption.Should().BeOfType<SymmetricDecryptionSettings>();
+            endpoint.Encryption.As<SymmetricDecryptionSettings>().Key.Should().BeNull();
+            endpoint.Encryption.As<SymmetricDecryptionSettings>().InitializationVector.Should().BeSameAs(iv);
+            endpoint.Encryption.As<SymmetricDecryptionSettings>().KeyProvider.Should().NotBeNull();
+            endpoint.Encryption.As<SymmetricDecryptionSettings>().KeyProvider!("abc").Should().BeSameAs(key);
         }
     }
 }