+++ docs/releases.md
@@ -4,6 +4,21 @@ uid: releases
 
 # Releases
 
+## [3.4.0-beta.1](https://github.com/BEagle1984/silverback/releases/tag/v3.4.0-beta.1)
+
+### What's new
+
+* Support encryption key rotation (see <xref:encryption>)
+
+# Releases
+
+## [3.3.1](https://github.com/BEagle1984/silverback/releases/tag/v3.3.1)
+
+### Fixes
+
+* Fix `AddHeaders<TMessage>` and `WithKafkaKey<TMessage>` not being correctly invoked by all `IProducer.Produce` and `IProducer.ProducerAsync` overloads
+* Add endpoint friendly name to all logs
+
 ## [3.3.0](https://github.com/BEagle1984/silverback/releases/tag/v3.3.0)
 
 ### What's new